+++ apps/web/src/routes/projects/$projectId/index.tsx
@@ -286,6 +286,8 @@ function ProjectDetailPage() {
                 <RepositoryAvatar
                   fullName={project.repository_full_name}
                   avatarUrl={project.repository_avatar_url}
+                  repositoryId={project.repository_id}
+                  provider={project.repository_provider}
                 />
                 {project.repository_full_name}
               </span>